+++ GUI/FunTrasGUI/FunTrasGUI/fun_tras.h
@@ -32,7 +32,7 @@ public:
     static cpp_dec_float_50 csc_t(cpp_dec_float_50 a);
     static cpp_dec_float_50 sec_t(cpp_dec_float_50 a);
     static cpp_dec_float_50 cot_t(cpp_dec_float_50 a);
-    static cpp_dec_float_50 log_t(float x, long y);
+    static cpp_dec_float_50 log_t(float x, float y);
     static cpp_dec_float_50 cosh_t(cpp_dec_float_50 a);
     static cpp_dec_float_50 sinh_t(cpp_dec_float_50 a);
     static cpp_dec_float_50 tanh_t(cpp_dec_float_50 a);
